# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Recetarium API",
    "version": "1.0.0",
    "description": "API REST para gestión de recetas personales. Permite crear, editar, eliminar y marcar recetas como favoritas."
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Desarrollo local"
    },
    {
      "url": "https://recetarium-one.vercel.app/api/v1",
      "description": "Producción (Vercel)"
    }
  ],
  "components": {
    "schemas": {
      "Ingrediente": {
        "type": "object",
        "required": [
          "nombre",
          "cantidad",
          "unidad",
          "familia"
        ],
        "properties": {
          "nombre": {
            "type": "string",
            "example": "Huevo"
          },
          "cantidad": {
            "type": "number",
            "example": 4
          },
          "unidad": {
            "type": "string",
            "example": "ud"
          },
          "familia": {
            "type": "string",
            "example": "lácteos"
          }
        }
      },
      "Receta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "a1b2c3d4-..."
          },
          "nombre": {
            "type": "string",
            "example": "Tortilla de patatas"
          },
          "categoria": {
            "type": "string",
            "example": "Española"
          },
          "sabor": {
            "type": "string",
            "enum": [
              "salado",
              "dulce",
              "amargo",
              "umami",
              "acido"
            ]
          },
          "tiempoPreparacion": {
            "type": "integer",
            "minimum": 1,
            "example": 30
          },
          "favorita": {
            "type": "boolean",
            "example": false
          },
          "imagen": {
            "type": "string",
            "format": "uri",
            "example": "https://images.unsplash.com/..."
          },
          "ingredientes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ingrediente"
            }
          },
          "pasos": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Batir los huevos"
            ]
          },
          "precioPorPorcion": {
            "type": "number",
            "example": 2.5
          },
          "porciones": {
            "type": "integer",
            "example": 2
          },
          "calorias": {
            "type": "integer",
            "example": 520
          },
          "proteinas": {
            "type": "number",
            "example": 38.5
          },
          "carbohidratos": {
            "type": "number",
            "example": 45
          },
          "grasas": {
            "type": "number",
            "example": 18
          },
          "hierro": {
            "type": "number",
            "example": 4.6,
            "readOnly": true,
            "description": "mg/porción. Lo calcula el servidor desde los ingredientes."
          },
          "sinGluten": {
            "type": "boolean",
            "nullable": true,
            "example": false,
            "readOnly": true,
            "description": "null si algún ingrediente no tiene ficha: ahí no se puede afirmar que no lleve."
          },
          "micros": {
            "type": "object",
            "readOnly": true,
            "description": "Micronutrientes por porción, calculados desde los ingredientes.",
            "properties": {
              "fibra": {
                "type": "number",
                "example": 6.2
              },
              "azucares": {
                "type": "number",
                "example": 8.1
              },
              "saturadas": {
                "type": "number",
                "example": 4.3
              },
              "sal": {
                "type": "number",
                "example": 1.94,
                "description": "La de los ingredientes; no cuenta la añadida al cocinar."
              },
              "hierroHemo": {
                "type": "number",
                "example": 1.8,
                "description": "Parte del hierro que es hemo (carne, pescado)."
              },
              "vitaminaC": {
                "type": "integer",
                "example": 42
              },
              "calcio": {
                "type": "integer",
                "example": 180
              },
              "b12": {
                "type": "number",
                "example": 1.4
              },
              "folato": {
                "type": "integer",
                "example": 96
              },
              "gluten": {
                "type": "object",
                "nullable": true,
                "properties": {
                  "fuentes": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "nombre": {
                          "type": "string",
                          "example": "salsa de soja"
                        },
                        "certeza": {
                          "type": "string",
                          "enum": [
                            "si",
                            "depende"
                          ]
                        },
                        "sustituto": {
                          "type": "string",
                          "nullable": true,
                          "example": "tamari sin gluten"
                        }
                      }
                    }
                  },
                  "evitable": {
                    "type": "boolean",
                    "example": true
                  }
                }
              },
              "estimadoDe": {
                "type": "string",
                "enum": [
                  "completo",
                  "parcial"
                ]
              }
            }
          },
          "tipo": {
            "type": "string",
            "enum": [
              "principal",
              "postre",
              "desayuno",
              "entrante"
            ],
            "example": "principal"
          }
        }
      },
      "RecetaInput": {
        "type": "object",
        "required": [
          "nombre",
          "sabor",
          "tiempoPreparacion",
          "ingredientes",
          "pasos"
        ],
        "properties": {
          "nombre": {
            "type": "string"
          },
          "categoria": {
            "type": "string"
          },
          "sabor": {
            "type": "string",
            "enum": [
              "salado",
              "dulce",
              "amargo",
              "umami",
              "acido"
            ]
          },
          "tiempoPreparacion": {
            "type": "integer",
            "minimum": 1
          },
          "imagen": {
            "type": "string",
            "format": "uri"
          },
          "ingredientes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ingrediente"
            }
          },
          "pasos": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "precioPorPorcion": {
            "type": "number"
          },
          "porciones": {
            "type": "integer"
          },
          "calorias": {
            "type": "integer"
          },
          "proteinas": {
            "type": "number"
          },
          "carbohidratos": {
            "type": "number"
          },
          "grasas": {
            "type": "number"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "principal",
              "postre",
              "desayuno",
              "entrante"
            ]
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "errores": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "paths": {
    "/despensa": {
      "get": {
        "summary": "Despensa del hogar (ingredientes en casa)",
        "tags": [
          "Plan"
        ],
        "responses": {
          "200": {
            "description": "Array de ingredientes con estado y caducidad opcional"
          }
        }
      },
      "put": {
        "summary": "Guardar la despensa del hogar",
        "tags": [
          "Plan"
        ],
        "responses": {
          "200": {
            "description": "Guardado"
          },
          "400": {
            "description": "Inválido"
          },
          "401": {
            "description": "Sesión requerida"
          }
        }
      }
    },
    "/plan": {
      "get": {
        "summary": "Obtener el plan semanal compartido",
        "tags": [
          "Plan"
        ],
        "responses": {
          "200": {
            "description": "Plan semanal (array de entradas)"
          }
        }
      },
      "put": {
        "summary": "Guardar el plan semanal compartido",
        "tags": [
          "Plan"
        ],
        "responses": {
          "200": {
            "description": "Plan guardado"
          },
          "400": {
            "description": "Plan inválido"
          },
          "401": {
            "description": "Sesión requerida"
          }
        }
      }
    },
    "/extras": {
      "get": {
        "summary": "Ítems manuales compartidos de la lista de la compra",
        "tags": [
          "Plan"
        ],
        "responses": {
          "200": {
            "description": "Array de ítems extra"
          }
        }
      },
      "put": {
        "summary": "Guardar los ítems manuales",
        "tags": [
          "Plan"
        ],
        "responses": {
          "200": {
            "description": "Guardado"
          },
          "400": {
            "description": "Inválido"
          },
          "401": {
            "description": "Sesión requerida"
          }
        }
      }
    },
    "/pendientes": {
      "get": {
        "summary": "Recetas compradas pendientes de planificar",
        "tags": [
          "Plan"
        ],
        "responses": {
          "200": {
            "description": "Array de recetas pendientes"
          }
        }
      },
      "put": {
        "summary": "Guardar las recetas pendientes de planificar",
        "tags": [
          "Plan"
        ],
        "responses": {
          "200": {
            "description": "Guardado"
          },
          "400": {
            "description": "Inválido"
          },
          "401": {
            "description": "Sesión requerida"
          }
        }
      }
    },
    "/recetas": {
      "get": {
        "summary": "Listar todas las recetas",
        "tags": [
          "Recetas"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "categoria",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por categoría"
          },
          {
            "in": "query",
            "name": "sabor",
            "schema": {
              "type": "string",
              "enum": [
                "salado",
                "dulce",
                "amargo",
                "umami",
                "acido"
              ]
            },
            "description": "Filtrar por sabor"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de recetas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Receta"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear una nueva receta",
        "tags": [
          "Recetas"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecetaInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Receta creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receta"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/recetas/{id}": {
      "get": {
        "summary": "Obtener una receta por ID",
        "tags": [
          "Recetas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Receta encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receta"
                }
              }
            }
          },
          "404": {
            "description": "Receta no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Actualizar una receta",
        "tags": [
          "Recetas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecetaInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Receta actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receta"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Receta no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Eliminar una receta (borrado lógico, se puede restaurar)",
        "tags": [
          "Recetas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminada correctamente"
          },
          "404": {
            "description": "Receta no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/recetas/{id}/favorita": {
      "patch": {
        "summary": "Alternar el estado de favorita",
        "tags": [
          "Recetas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Receta con favorita actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receta"
                }
              }
            }
          },
          "404": {
            "description": "Receta no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/recetas/{id}/restaurar": {
      "post": {
        "summary": "Devolver al catálogo una receta borrada, con su mismo id",
        "tags": [
          "Recetas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Receta restaurada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no estaba borrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/admin/sesiones": {
      "get": {
        "summary": "Quién ha iniciado sesión, desde qué IP y con qué dispositivo",
        "description": "Solo administradores. Devuelve sesiones recientes, resumen por usuario e IPs vistas por primera vez.",
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limite",
            "schema": {
              "type": "integer",
              "default": 100,
              "maximum": 500
            }
          },
          {
            "in": "query",
            "name": "dias",
            "schema": {
              "type": "integer",
              "default": 7,
              "maximum": 90
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sesiones",
            "resumen por usuario e IPs nuevas": null
          },
          "401": {
            "description": "Sesión requerida"
          },
          "403": {
            "description": "Solo para administradores"
          }
        }
      }
    },
    "/admin/invitados": {
      "get": {
        "summary": "Lista blanca de acceso, y los hogares disponibles",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Invitaciones y hogares"
          },
          "403": {
            "description": "Solo para administradores"
          }
        }
      },
      "post": {
        "summary": "Invitar un correo",
        "description": "hogarId nulo crea un hogar propio en el primer inicio de sesión.",
        "tags": [
          "Admin"
        ],
        "responses": {
          "201": {
            "description": "Invitación guardada"
          },
          "400": {
            "description": "Correo",
            "rol u hogar inválidos": null
          },
          "403": {
            "description": "Solo para administradores"
          }
        }
      }
    },
    "/admin/invitados/{email}": {
      "delete": {
        "summary": "Retirar una invitación que nadie haya usado",
        "tags": [
          "Admin"
        ],
        "responses": {
          "204": {
            "description": "Retirada"
          },
          "403": {
            "description": "Solo para administradores"
          },
          "404": {
            "description": "No hay invitación pendiente para ese correo"
          }
        }
      }
    },
    "/yo": {
      "get": {
        "summary": "Quién soy, según el token de sesión",
        "description": "Fuente de verdad de la sesión al arrancar la app. En la PWA de iOS la cookie del servicio de auth no sobrevive al cierre, así que el cliente guarda el token y pregunta aquí: este backend valida contra neon_auth.session, que está en su misma base de datos.\n",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Usuario",
            "hogar y rol": null
          },
          "401": {
            "description": "Sesión inválida o caducada"
          },
          "403": {
            "description": "Sin acceso"
          }
        }
      }
    },
    "/yo/sesion": {
      "delete": {
        "summary": "Cerrar sesión",
        "description": "Borra las sesiones del usuario, así que la revocación es inmediata. Se cierran todas y no solo la de este dispositivo, porque el cliente manda un JWT y no el identificador de la fila.\n",
        "tags": [
          "Auth"
        ],
        "responses": {
          "204": {
            "description": "Sesión cerrada"
          },
          "401": {
            "description": "Sesión inválida o caducada"
          }
        }
      }
    }
  },
  "tags": []
}